# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/payments":{"get":{"summary":"Get Payments","operationId":"get_payments_api_payments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/school/profile":{"get":{"summary":"Get School Profile","operationId":"get_school_profile_api_school_profile_get","security":[{"HTTPBearer":[]}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"patch":{"summary":"Update School Profile","operationId":"update_school_profile_api_school_profile_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string","title":"Name"}},{"name":"phone","in":"query","required":false,"schema":{"type":"string","title":"Phone"}},{"name":"address","in":"query","required":false,"schema":{"type":"string","title":"Address"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_update_school_profile_api_school_profile_patch"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/school/me":{"get":{"summary":"Get My School","operationId":"get_my_school_api_school_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/dashboard/stats":{"get":{"summary":"Get Dashboard Stats","operationId":"get_dashboard_stats_api_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/pending-users":{"get":{"summary":"Get Pending Users","operationId":"get_pending_users_api_admin_pending_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/staff":{"get":{"summary":"Get Staff","operationId":"get_staff_api_staff_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/pending":{"get":{"summary":"Get All Pending","operationId":"get_all_pending_api_admin_pending_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/approve/{item_type}/{item_id}":{"patch":{"summary":"Approve Item","operationId":"admin_approve_item_action","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_type","in":"path","required":true,"schema":{"type":"string","title":"Item Type"}},{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalActionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register-school":{"post":{"summary":"Register School","operationId":"register_school_api_auth_register_school_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterSchoolRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/join-school":{"post":{"summary":"Join School","operationId":"join_school_api_auth_join_school_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Request"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"summary":"Login","operationId":"auth_login_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/school/invite":{"get":{"summary":"Get School Invite","operationId":"get_school_invite_api_school_invite_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/schools":{"get":{"summary":"Get Schools","operationId":"get_schools_api_schools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/schools/{school_id}":{"get":{"summary":"Get School","operationId":"get_school_api_schools__school_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"school_id","in":"path","required":true,"schema":{"type":"string","title":"School Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update School","operationId":"update_school_api_schools__school_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"school_id","in":"path","required":true,"schema":{"type":"string","title":"School Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Update Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/students":{"post":{"summary":"Create Student","operationId":"create_student_api_students_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudentRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Students","operationId":"get_students_api_students_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"approval_status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approval Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/students/{student_id}":{"get":{"summary":"Get Student","operationId":"get_student_api_students__student_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Student","operationId":"update_student_api_students__student_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Update Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/students/{student_id}/status":{"patch":{"summary":"Update Student Status","operationId":"update_student_status_api_students__student_id__status_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/payments/initiate":{"post":{"summary":"Initiate Payment","operationId":"initiate_payment_api_payments_initiate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InitiatePaymentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/attendance":{"post":{"summary":"Mark Attendance","operationId":"mark_attendance_api_attendance_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarkAttendanceRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Attendance","operationId":"get_attendance_api_attendance_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date"}},{"name":"approval_status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approval Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/results":{"post":{"summary":"Record Result","operationId":"record_result_api_results_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordResultRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/results/{student_id}":{"get":{"summary":"Get Student Results","operationId":"get_student_results_api_results__student_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/announcements":{"get":{"summary":"Get Announcements","operationId":"get_announcements_api_announcements_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"approval_status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approval Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/finance/request-approval":{"post":{"summary":"Request Admin Approval","operationId":"finance_request_admin_approval","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Request"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/approval-requests":{"get":{"summary":"Get Approval Requests","operationId":"get_approval_requests_api_admin_approval_requests_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/webhooks/mpesa/callback":{"post":{"summary":"Mpesa Callback","operationId":"mpesa_callback_api_webhooks_mpesa_callback_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/portal/my-data":{"get":{"summary":"Get My Portal Data","operationId":"get_my_portal_data_api_portal_my_data_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/finance/transactions":{"get":{"summary":"Get Transactions","operationId":"get_transactions_api_finance_transactions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create Transaction","operationId":"create_transaction_api_finance_transactions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/finance/summary":{"get":{"summary":"Get Finance Summary","operationId":"get_finance_summary_api_finance_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/progress-students":{"post":{"summary":"Progress Students","operationId":"progress_students_api_admin_progress_students_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProgressStudentsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/students/{student_id}/history":{"get":{"summary":"Get Student History","operationId":"get_student_history_api_students__student_id__history_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApprovalActionRequest":{"properties":{"action":{"type":"string","title":"Action"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"}},"type":"object","required":["action"],"title":"ApprovalActionRequest"},"AttendanceStatus":{"type":"string","enum":["present","absent","late"],"title":"AttendanceStatus"},"Body_update_school_profile_api_school_profile_patch":{"properties":{"logo":{"type":"string","contentMediaType":"application/octet-stream","title":"Logo"}},"type":"object","title":"Body_update_school_profile_api_school_profile_patch"},"CreateStudentRequest":{"properties":{"admission_number":{"type":"string","title":"Admission Number"},"full_name":{"type":"string","title":"Full Name"},"date_of_birth":{"type":"string","title":"Date Of Birth"},"gender":{"type":"string","title":"Gender"},"class_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Class Name"},"year_of_study":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Year Of Study"},"stream":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stream"},"guardian_name":{"type":"string","title":"Guardian Name"},"guardian_phone":{"type":"string","title":"Guardian Phone"},"guardian_email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Guardian Email"},"guardian_relationship":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Guardian Relationship"},"secondary_guardian_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Secondary Guardian Name"},"secondary_guardian_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Secondary Guardian Phone"},"secondary_guardian_email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Secondary Guardian Email"},"secondary_guardian_relationship":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Secondary Guardian Relationship"},"blood_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blood Type"},"allergies":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allergies"},"chronic_conditions":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Chronic Conditions"},"disabilities":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Disabilities"},"immunization_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Immunization Status"},"medical_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medical Info"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","default":"active"}},"type":"object","required":["admission_number","full_name","date_of_birth","gender","guardian_name","guardian_phone"],"title":"CreateStudentRequest"},"CreateTransactionRequest":{"properties":{"transaction_type":{"type":"string","title":"Transaction Type"},"category":{"type":"string","title":"Category"},"amount":{"type":"number","title":"Amount"},"description":{"type":"string","title":"Description"},"date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date"}},"type":"object","required":["transaction_type","category","amount","description"],"title":"CreateTransactionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InitiatePaymentRequest":{"properties":{"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"amount":{"type":"number","title":"Amount"},"payment_type":{"type":"string","title":"Payment Type"},"payment_method":{"type":"string","title":"Payment Method"},"student_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Id"},"bank_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bank Reference"},"cheque_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cheque Number"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["amount","payment_type","payment_method"],"title":"InitiatePaymentRequest"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MarkAttendanceRequest":{"properties":{"entity_type":{"type":"string","title":"Entity Type"},"entity_id":{"type":"string","title":"Entity Id"},"date":{"type":"string","title":"Date"},"status":{"$ref":"#/components/schemas/AttendanceStatus"},"remarks":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remarks"}},"type":"object","required":["entity_type","entity_id","date","status"],"title":"MarkAttendanceRequest"},"ProgressStudentsRequest":{"properties":{"academic_year":{"type":"string","title":"Academic Year"},"from_class":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Class"}},"type":"object","required":["academic_year"],"title":"ProgressStudentsRequest"},"RecordResultRequest":{"properties":{"exam_id":{"type":"string","title":"Exam Id"},"student_id":{"type":"string","title":"Student Id"},"subject":{"type":"string","title":"Subject"},"marks":{"type":"number","title":"Marks"},"grade":{"type":"string","title":"Grade"},"teacher_comments":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Teacher Comments"}},"type":"object","required":["exam_id","student_id","subject","marks","grade"],"title":"RecordResultRequest"},"RegisterSchoolRequest":{"properties":{"name":{"type":"string","title":"Name"},"address":{"type":"string","title":"Address"},"phone":{"type":"string","title":"Phone"},"email":{"type":"string","format":"email","title":"Email"},"school_type":{"type":"string","title":"School Type"},"logo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logo Url"},"motto":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motto"},"vision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vision"},"mission":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mission"},"principal_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Principal Name"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"established_year":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Established Year"},"admin_name":{"type":"string","title":"Admin Name"},"admin_email":{"type":"string","format":"email","title":"Admin Email"},"admin_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Admin Phone"},"admin_password":{"type":"string","title":"Admin Password"}},"type":"object","required":["name","address","phone","email","school_type","admin_name","admin_email","admin_password"],"title":"RegisterSchoolRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}